Create a new notification window for each build run
This removes the global variable `currentNotification` and replaces it with a local reference cell that is passed to all functions that report notifications.

This means that when you start a build, hit `Esc` to hide the notification window and then start a _new_ build, the old build will not add notifications to the new window - so the new notification window only shows the new notifications.